Refactor OrderService and add target method to Calculate
The OrderService class was refactored to provide additional flexibility when opening an order. Options for take-profit, stop-loss, and leverage were added and the method was renamed to "open" from "long". Additionally, a "target" method was implemented in the Calculate class to provide a way to calculate the target price based on an input percentage.