Add asset balance functionality to the exchange service
A new function for retrieving futures balance data has been added to the Exchange Service, along with the necessary interfaces and classes. This includes a new contract, AssetBalanceContract, and a new method in the CoinexService class. A new model for Asset and a collection model, AssetCollection, have also been added for managing the data.